+++ src/main/java/date/Date.java
@@ -1,0 +1,22 @@
+package date;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public abstract class Date {
+    private LocalDate date;
+
+    public Date(String date) throws DateTimeParseException {
+        this.date = LocalDate.parse(date);
+    }
+
+    @Override
+    public String toString() {
+        return date.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
+    }
+
+    public String encode() {
+        return date.toString();
+    }
+}
